LayoutTests/media/modern-media-controls/media-documents/ipad/media-document-video-ios-sizing-expected.txt

Testing the size of the media element in a video media document on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getComputedStyle(media).width became "700px"
PASS getComputedStyle(media).height is "525px"

PASS successfullyParsed is true

TEST COMPLETE
